move prepare_earth_position_vel
